+++ test/Fakes/CleanRegex/Internal/Model/IgnoreAware.php
@@ -1,0 +1,12 @@
+<?php
+namespace Test\Fakes\CleanRegex\Internal\Model;
+
+use TRegx\CleanRegex\Internal\Model\GroupHasAware;
+
+class IgnoreAware implements GroupHasAware
+{
+    public function hasGroup($nameOrIndex): bool
+    {
+        return true;
+    }
+}